Log Guidelines
Proposed final draft.

This attempts to integrate much of the existing feedback, including
clarity around the points that were confusing people regarding the
"unit of work" and when stacktraces are appropriate.

In addition the Security logging section was dropped, because
realistically it was a poor fit for this original document. The wiki
page on it remains. My expectation is that security for logging should
be proposed as a follow on patch by someone that wants to make that
info fit well with this document.

All the open questions are now removed. Hopefully this ends up very
close to an agreed point.

Note: This evolved originally in the nova-specs repo at:
I674966988ed501dee323a106b84b5b6602d846b1 but is now moved over to
cross project specs repo. Some deep history can be seen at the
original draft.

Part of bp:log-guidelines

Change-Id: I3a8d0ac570e05eb11ccc70ad4eb9da277aaac256